# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SignalGrid Device Policy & Security API",
    "description": "Enterprise device security, compliance, and policy enforcement platform. Integrates with MDM, NAC, ITSM, and SIEM systems.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@signalgrid.local"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.signalgrid.local",
      "description": "Production API",
      "variables": {}
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/location/report": {
      "post": {
        "summary": "Report device location",
        "description": "Submit location signal from device or location service",
        "operationId": "reportLocation",
        "tags": ["Telemetry"],
        "security": [
          {
            "HMACSignature": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationSignal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "summary": "List security events",
        "description": "Get audit log of security decisions and policy actions",
        "operationId": "listEvents",
        "tags": ["Events"],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by event type",
            "schema": {
              "type": "string",
              "enum": ["session_allowed", "session_denied", "quarantine", "siem_alert", "itsm_ticket"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event list with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from OIDC provider or admin API key"
      },
      "APIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Admin API key for development (deprecated in production)"
      },
      "HMACSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "HMAC-SHA256 request signature for device endpoints"
      }
    },
    "schemas": {
...
      "LocationSignal": {
        "type": "object",
        "required": ["deviceId", "observedAt", "source"],
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string",
            "enum": ["gps", "wifi", "ble", "cellular", "manual"]
          },
          "mode": {
            "type": "string",
            "enum": ["precise", "coarse", "presence"]
          },
          "lat": {
            "type": "number"
          },
          "lon": {
            "type": "number"
          },
          "accuracyM": {
            "type": "number",
            "description": "Accuracy in meters"
          },
          "siteId": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "floorId": {
            "type": "string"
          },
          "zoneId": {
            "type": "string"
          }
        }
      },
      "SecurityEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["session_denied", "session_allowed", "quarantine", "siem_alert", "itsm_ticket"]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "actor": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "device": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "complianceStatus": {
                "type": "string"
              }
            }
          },
          "decision": {
            "type": "string",
            "enum": ["DENY", "ALLOW"]
          },
          "reason": {
            "type": "string"
          },
          "actionsTriggered": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "riskScore": {
            "type": "number"
          }
        }
      },
      "EventList": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityEvent"
            }
          },
          "total": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimitError": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}